refactor: remove type.ts, add widget-config-list.ts
