+++ src/services/dashboards/widgets/_base/base-trend/widget-config.ts
@@ -1,3 +1,18 @@
-import type { WidgetConfig } from '@/services/dashboards/widgets/type';
+import type { WidgetConfig } from '@/services/dashboards/widgets/config';
 
-export default {} as WidgetConfig;
+const baseTrendWidgetConfig: WidgetConfig = {
+    widget_config_id: 'baseTrend',
+    widget_component: () => ({
+        component: import('@/services/dashboards/widgets/_base/base-trend/BaseTrendWidget.vue'),
+    }),
+    labels: ['Cost'],
+    description: {
+    },
+    scopes: ['PROJECT', 'WORKSPACE'],
+    theme: {
+        inherit: true,
+    },
+    sizes: ['lg', 'full'],
+};
+
+export default baseTrendWidgetConfig;
