+++ src/services/dashboards/widgets/aws-data-transfer-cost-trend/widget-config.ts
@@ -1,3 +1,3 @@
-import type { WidgetConfig } from '@/services/dashboards/widgets/type';
+import type { WidgetConfig } from '@/services/dashboards/widgets/config';
 
 export default {} as WidgetConfig;
