+++ src/services/dashboards/widgets/cost-trend/widget-config.ts
@@ -1,3 +1,16 @@
-import type { WidgetConfig } from '@/services/dashboards/widgets/type';
+import type { WidgetConfig } from '@/services/dashboards/widgets/config';
 
-export default {} as WidgetConfig;
+const costTrendWidgetConfig: WidgetConfig = {
+    widget_config_id: 'costTrend',
+    base_configs: [{ config_id: 'baseTrend' }],
+    labels: ['Cost'],
+    description: {
+    },
+    scopes: ['PROJECT', 'WORKSPACE'],
+    theme: {
+        inherit: true,
+    },
+    sizes: ['lg', 'full'],
+};
+
+export default costTrendWidgetConfig;
